修改brdf；重构resource mgr

+++ Engine/Inc/Framework/ImGui/Widgets/CommonTextureWidget.h
@@ -8,17 +8,17 @@ namespace Ailu
 	class TextureSelector : public ImGuiWidget
 	{
 	public:
-		inline static constexpr u64 kInvalidTextureID = std::numeric_limits<u64>::max();
-		static bool IsValidTextureID(u64 id) { return id != kInvalidTextureID; }
+		inline static constexpr u32 kInvalidTextureID = std::numeric_limits<u32>::max();
+		static bool IsValidTextureID(u32 id) { return id != kInvalidTextureID; }
 		TextureSelector();
 		~TextureSelector();
 		void Open(const i32& handle) final;
 		void Close(i32 handle) final;
-		u64 GetSelectedTexture(i32 handle) const;
+		u32 GetSelectedTexture(i32 handle) const;
 	private:
 		void ShowImpl() final;
 		i32 _selected_img_index = -1;
-		u64 _cur_selected_texture_id = kInvalidTextureID;
+		u32 _cur_selected_texture_id = kInvalidTextureID;
 	};
 
 	class RenderTextureView : public ImGuiWidget
